+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.3.0 (Unreleased)
 
--- Modify the parsing syntax:
+- Modify the parsing syntax:
   1. Stop allowing abstractions without delimiters, *e.g.* `λx x` is not allowed
      anymore, use `λx. x` or `λx -> x` instead.
   2. Allow spaces between the lambda symbol and the variable, *e.g.* `λ x -> x`
@@ -8,10 +8,16 @@
       documentation, but the implementation was not consistent.
   3. Allow inner abstractions to be omitted, *e.g.* `λx. λy. x y` can be
      written as `λx y. x y`.
+  4. Stop allowing special depth variables such as `_x1` or `_y2`. These
+     underscore variables are not prohibited.
 
--- Fix bugs in the parser.
+- Improve the pretty-printing of `Lambda` to not creating variables with
+  underscores. Instead it will choose appropriate fresh names that are
+  guaranteed to not conflict with existing variables.
 
--- More documentation.
+- Fix bugs in the parser.
+
+- More documentation.
 
 ## 1.2.0